requirements.txt
win32crypt
pycryptodome
